Fix mobile navigation
I believe the intention of the forced horizontal flex row was to
make the social buttons always be side-by-side but the rules also
affected how the main menu is laid out in mobile.

Now the rules only apply to the social buttons which keeps the same
behavior as before for them but doesn't break navigation.